added rolls in UserController

+++ app/Http/Controllers/UserController.php
@@ -16,7 +16,16 @@ class UserController extends Controller
      */
     public function index()
     {
-        return User::all();
+        $user = Auth::guard('sanctum')->user();
+
+        if($user->isAbleTo(['users-read']))
+        {
+            return $this->sendResponse(User::all(), 'Successfully pulled all users.');
+        }
+        else
+        {
+            return $this->sendError('Unauthorised.');
+        }
     }
 
     /**
@@ -27,16 +36,25 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
-        $user = new User();
-        $user->fill($request->toArray());
-        $user->password = bcrypt(request('password'));
+        $user = Auth::guard('sanctum')->user();
 
-        $user->save();
-        $user->attachRole($request->role);
+        if($user->isAbleTo(['users-create']))
+        {
+            $user = new User();
+            $user->fill($request->toArray());
+            $user->password = bcrypt(request('password'));
 
-        $token = $user->createToken('ApiAuth')->plainTextToken;
+            $user->save();
+            $user->attachRole($request->role);
 
-        return $this->sendResponse(['token' => $token],$user, 'Successfully registered.');
+            //$token = $user->createToken('ApiAuth')->plainTextToken;
+            //['token' => $token],
+            return $this->sendResponse($user, 'Successfully registered.');
+        }
+        else
+        {
+            return $this->sendError('Unauthorised.');
+        }
     }
 
     /**
@@ -47,7 +65,25 @@ class UserController extends Controller
      */
     public function show(User $user)
     {
-        return $user;
+        $user = Auth::guard('sanctum')->user();
+
+        if($user->isAbleTo(['users-read']))
+        {
+            $user = User::find($user);
+            if($user === null)
+            {
+                return $this->sendError('Empty or does not exist.');
+            }
+            else
+            {
+                return $this->sendResponse(['user' => $user], 'User has been successfully pulled');
+
+            }
+        }
+        else
+        {
+            return $this->sendError('Unauthorised.');
+        }
     }
 
     /**
@@ -59,9 +95,33 @@ class UserController extends Controller
      */
     public function update(Request $request, User $user)
     {
-        $user->fill($request->toArray());
-
-        return $user->save();
+        $user_data = Auth::guard('sanctum')->user();
+        
+        if($user_data->isAbleTo(['users-update']))
+        {
+            $user->fill($request->toArray());
+            $user->save();
+            if($user_data->isAbleTo(['role-update']))
+            {
+                $rolls = explode(",",$request->role);
+                if($rolls == !""){
+                    $user->detachRoles($user->roles);
+                }
+                if(is_array($rolls)){
+                    foreach($rolls as $rol){
+                        $user->attachRole($rol);
+                    }
+                }
+                else{
+                    $user->attachRole($rolls);
+                }
+            }
+            return $this->sendResponse([], 'User has been successfully updated');
+        }
+        else
+        {
+            return $this->sendError('Unauthorised.');
+        }
     }
 
     /**
@@ -72,6 +132,17 @@ class UserController extends Controller
      */
     public function destroy(User $user)
     {
-        return User::destroy($user->id);
+        $user = Auth::guard('sanctum')->user();
+
+        if($user->isAbleTo(['users-delete']))
+        {
+            User::destroy($user->id);
+
+            return $this->sendResponse('Successfully deleted.');
+        }
+        else
+        {
+            return $this->sendError('Unauthorised.');
+        }
     }
 }
